chore: rename package `member` to `auth`
