+++ src/main/kotlin/ecommerce/auth/data/TokenRequest.kt
@@ -1,0 +1,10 @@
+package ecommerce.auth.data
+
+data class TokenRequest(
+//    @field:Size()
+//    @field:Pattern()
+    val email: String,
+//    @field:Size()
+//    @field:Pattern()
+    val password: String,
+)
