+++ src/main/kotlin/ecommerce/auth/data/TokenResponse.kt
@@ -1,0 +1,5 @@
+package ecommerce.auth.data
+
+data class TokenResponse(
+    val accessToken: String,
+)
